Refaktorer pakker og optimaliserer tester.
